fix: initial highlight with null value
